+++ adapters/heiman/__init__.py
@@ -4,12 +4,12 @@ from adapters.generic.smoke_sensor import SmokeSensorAdapter
 from adapters.generic.temp_hum_sensor import TemperatureHumiditySensorAdapter as TempHumAdapter
 from adapters.generic.water_leak_sensor import WaterLeakSensorAdapter
 from adapters.generic.on_off_kwh import OnOffKwhAdapter
+from adapters.generic.siren import SirenAdapterWithBattery
 from adapters.on_off_switch_adapter import OnOffSwitchAdapter
 from adapters.heiman.gas_sensor import HeimanGasSensorAdapter as GasAdapter
 from adapters.heiman.HS1CAE import HS1CAE
 from adapters.heiman.HS1WLE import HS1WLE
 from adapters.heiman.HS2SK import HS2SK
-from adapters.heiman.HS2WDE import HS2WDE
 from adapters.heiman.SMHMI1 import SMHMI1
 
 
@@ -26,7 +26,7 @@ heiman_adapters = {
     'HS1WL': WaterLeakSensorAdapter,    # HEIMAN Water leakage sensor
     'HS2ESK-E': OnOffKwhAdapter,        # HEIMAN Smart in wall plug
     'HS2SK': HS2SK,                     # HEIMAN Smart metering plug
-    'HS2WD-E': HS2WDE,                  # HEIMAN Smart siren
+    'HS2WD-E': SirenAdapterWithBattery, # HEIMAN Smart siren
     'HS3MS': MotionSensorAdapter,       # HEIMAN Smart motion sensor
     'HS3SA': SmokeSensorAdapter,        # HEIMAN Smoke detector
     'SKHMP30-I1': HS2SK,                # HEIMAN Smart metering plug
@@ -35,6 +35,7 @@ heiman_adapters = {
     'STHM-I1H': TempHumAdapter,         # HEIMAN temperature & humidity sensor
     'SWHM-I1': WaterLeakSensorAdapter,  # HEIMAN Water leakage sensor
     'HS3CG': GasAdapter,                # HEIMAN Combustible gas sensor
+    'HS1CG-E': GasAdapter,              # HEIMAN Combustible gas sensor
     'HS1CG-M': GasAdapter,              # HEIMAN Combustible gas sensor
     'HS1CG_M': GasAdapter,              # HEIMAN Combustible gas sensor
     'SGMHM-I1': GasAdapter,             # HEIMAN Combustible gas sensor